feat: Add form submission functionality for comments
This commit adds JavaScript code to handle form submission for comments on the website. When the comment form is submitted, the data is collected and sent to a backend endpoint using a POST request. If the request is successful, a success message is displayed to the user. If there is an error, an error message is displayed.

The code changes include:
- Adding event listener to the comment form submit event
- Collecting form data and creating a JSON object
- Sending a POST request to the backend endpoint with the form data
- Handling the response and displaying appropriate messages to the user